+++ ICalculadora.java
@@ -1,5 +1,5 @@
 /**
- * Interfaz que define operaciones básicas de una calculadora.
+ * doubleerfaz que define operaciones básicas de una calculadora.
  */
 public interface ICalculadora {
 
@@ -9,7 +9,7 @@ public interface ICalculadora {
      * @param op2 El segundo operando.
      * @return El resultado de la suma.
      */
-    double sumar(int op1, int op2);
+    double sumar(double op1, double op2);
 
     /**
      * Realiza la operación de resta entre dos operandos.
@@ -17,7 +17,7 @@ public interface ICalculadora {
      * @param op2 El segundo operando.
      * @return El resultado de la resta.
      */
-    double restar(int op1, int op2);
+    double restar(double op1, double op2);
 
     /**
      * Realiza la operación de multiplicación entre dos operandos.
@@ -25,14 +25,14 @@ public interface ICalculadora {
      * @param op2 El segundo operando.
      * @return El resultado de la multiplicación.
      */
-    double multiplicar(int op1, int op2);
+    double multiplicar(double op1, double op2);
 
     /**
      * Realiza la operación de división entre dos operandos.
      * @param op1 El numerador.
      * @param op2 El denominador.
      * @return El resultado de la división.
-     * @throws ArithmeticException Si se intenta dividir por cero.
+     * @throws ArithmeticException Si se doubleenta dividir por cero.
      */
-    double dividir(int op1, int op2) throws ArithmeticException;
+    double dividir(double op1, double op2) throws ArithmeticException;
 }
